Add timestamped sample buffer
When a sample is received, it will now be stored in a buffer. The buffer stores
1 second's worth of samples, with the start of the buffer aligned to the start
of the second on the clock. Since neither 9-2 LE nor 61869-9 provide a full
timestamp for each sample (only the sub-second part is provided through the
smpCnt field), we assume that it was sampled in the same second that it was
received, unless that time would be in the future, in which case we assume that
it was for the previous second. This, of course, assumes that the system clock
is synchronised with that of the merging unit.

Currently no functionality is present for sending this data onwards; it simply
counts the number of non-zero values and prints the results for debugging
purposes before flushing the buffer.