Add XML sample buffer output
The sample buffer now outputs data as OpenPMU XML datagrams. Since the OpenPMU
documentation states that these datagrams are sent at twice the nominal
frequency, the sample buffer can now hold less than a full second of samples.
In the future, the XML-based format will likely be replaced with a JSON-based
one, but the code should be easy enough to adapt.

Currently, the sample buffer is flushed as soon as a sample is received for the
next buffer. This means that some samples may be lost towards the end of each
frame if they are received out of order. To fix this, we would need to keep
each buffer around for a short time after the next one is started.