Retain two sample buffers simultaneously
Previously, only one sample buffer was used, and was flushed as soon as a
sample which fell outside of its timespan was received. This meant that the
last few samples of each buffer would be lost if they were received out of
order. Now, we hold on to the previous buffer as well as the current one, only
flushing it once another new buffer is created (at which point the current
buffer becomes the previous buffer).