Add buffer queue, part 1
This commit replaces the 'current' and 'previous' buffers with a proper queue.
This gives us greater control over latency, since a buffer no longer has to
wait for the next buffer to be filled before it can be sent.

The documentation hasn't yet been updated to accommodate this change, and for
some reason the buffer send times are off by a second, hence 'part 1'.